Moved __parts*
