+++ aksel.nav.no/website/components/website-modules/examples/__parts/dekorator.tsx
